requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
gunicorn==21.2.0
sentence-transformers==2.7.0
torch==2.3.0
numpy==1.26.4
typesense==0.21.0

# LangChain and Groq LLM
langchain==0.1.20
langchain-groq==0.1.3
langchain-community==0.0.38
tiktoken==0.6.0